+++ tests/admin/test_base.py
@@ -65,61 +65,10 @@ def test_baseview_defaults():
     assert view.endpoint == "mockview"
     assert view.url is None
     assert view.static_folder is None
-    assert view.category is None
     assert view.admin is None
     assert view.blueprint is None
 
-
-def test_admin_category():
-    admin = Admin()
-    admin.add_category("Category1", "class-name", "icon-type", "icon-value")
-    view_1 = MockView(name="Test 1", endpoint="test1", category="Category1")
-    view_2 = MockView(name="Test 2", endpoint="test2", category="Category2")
-    view_3 = MockView(name="Test 3", endpoint="test3", category="Category2")
-
-    admin.add_view(view_1)
-    admin.add_view(view_2)
-    admin.add_view(view_3)
-
-    # print(admin.menu())
-    # print(admin._menu_categories)
-    # print(admin._menu_links)
-
-    assert "Category1" in admin._menu_categories
-    assert "Category2" in admin._menu_categories
-
-    for menu in admin.menu():
-        if menu.name == "Category1":
-            menu_category1 = menu
-        if menu.name == "Category2":
-            menu_category2 = menu
-
-    assert menu_category1.get_class_name() == "class-name"
-    assert menu_category1.get_icon_type() == "icon-type"
-    assert menu_category1.get_icon_value() == "icon-value"
-    assert len(menu_category1.get_children()) == 1
-    assert menu_category1.get_children()[0].name == "Test 1"
-
-    assert menu_category2.get_class_name() is None
-    assert menu_category2.get_icon_type() is None
-    assert menu_category2.get_icon_value() is None
-    assert len(menu_category2.get_children()) == 2
-    assert menu_category2.get_children()[0].name == "Test 2"
-    assert menu_category2.get_children()[1].name == "Test 3"
-
-    # Categories don't have URLs
-    assert menu_category1.get_url() is None
-    assert menu_category2.get_url() is None
-
-    view_3.allow_access = False
-    # Categories are only accessible if there is at least one accessible child
-    assert menu_category2.is_accessible()
-    children = menu_category2.get_children()
-    assert len(children) == 1
-    assert children[0].is_accessible()
-
-
-def test_admin_defaults(app, client, admin):
+def test_admin_default(app, client, admin):
     assert "admin" in app.extensions
     assert len(app.extensions["admin"]) == 1
 
@@ -140,20 +89,20 @@ def test_admin_defaults(app, client, admin):
 
     # Check if default view was added
     assert len(admin._views) == 2
-    assert admin._views[0] == admin.index_view
-    assert admin._views[1] == admin.user_view
-
+    index_view = admin._views[0]
+    user_view = admin._views[1]
+    
     # check index_view
-    assert admin.index_view is not None
-    assert admin.index_view.endpoint == "index"
-    assert admin.index_view.url == "/"
-    assert admin.index_view.index_template == "admin/index.html"
+    assert index_view is not None
+    assert index_view.endpoint == "index"
+    assert index_view.url == "/"
+    assert index_view.index_template == "index.html"
 
     # check user_view
-    assert admin.user_view is not None
-    assert admin.user_view.endpoint == "user"
-    assert admin.user_view.url == "/user"
-    assert admin.user_view.index_template == "views/user/index.html"
+    assert user_view is not None
+    assert user_view.endpoint == "user"
+    assert user_view.url == "/user"
+    assert user_view.index_template == "views/user/index.html"
 
     assert get_app_endpoint_rule(app, "index.index") == "/"
     assert get_app_endpoint_rule(app, "index.admin_index") == "/admin/"
@@ -189,6 +138,57 @@ def test_admin_defaults(app, client, admin):
     assert rv.status_code == 200
     rv = client.get(user_logout_url)
     assert rv.status_code == 302
+
+
+def test_admin_menu():
+    admin = Admin()
+    menu=admin.menu
+    assert menu.admin == admin
+    menu.add_category("Category1", "class-name", "icon-type", "icon-value")
+    view_1 = MockView(name="Test 1", endpoint="test1")
+    view_2 = MockView(name="Test 2", endpoint="test2")
+    view_3 = MockView(name="Test 3", endpoint="test3")
+
+    admin.add_view(view_1, category="Category1")
+    admin.add_view(view_2, category="Category2")
+    admin.add_view(view_3, category="Category2")
+
+    # print(menu._menu)
+    # print(menu._menu_categories)
+    # print(menu._menu_links)
+
+    assert "Category1" in menu._menu_categories
+    assert "Category2" in menu._menu_categories
+
+    for m in menu.menus():
+        if m.name == "Category1":
+            menu_category1 = m
+        if m.name == "Category2":
+            menu_category2 = m
+
+    assert menu_category1.get_class_name() == "class-name"
+    assert menu_category1.get_icon_type() == "icon-type"
+    assert menu_category1.get_icon_value() == "icon-value"
+    assert len(menu_category1.get_children()) == 1
+    assert menu_category1.get_children()[0].name == "Test 1"
+
+    assert menu_category2.get_class_name() is None
+    assert menu_category2.get_icon_type() is None
+    assert menu_category2.get_icon_value() is None
+    assert len(menu_category2.get_children()) == 2
+    assert menu_category2.get_children()[0].name == "Test 2"
+    assert menu_category2.get_children()[1].name == "Test 3"
+
+    # Categories don't have URLs
+    assert menu_category1.get_url() is None
+    assert menu_category2.get_url() is None
+
+    view_3.allow_access = False
+    # Categories are only accessible if there is at least one accessible child
+    assert menu_category2.is_accessible()
+    children = menu_category2.get_children()
+    assert len(children) == 1
+    assert children[0].is_accessible()
 
 
 def test_admin_init_app_raise(app):
@@ -239,8 +239,9 @@ def test_inaccessible_callback(client, admin):
 
 
 def test_menu_links(client, admin):
-    admin.add_link(MenuLink("TestMenuLink1", endpoint=".index"))
-    admin.add_link(MenuLink("TestMenuLink2", url="http://python.org/"))
+    menu = admin.menu
+    menu.add_link(MenuLink("TestMenuLink1", endpoint=".index"))
+    menu.add_link(MenuLink("TestMenuLink2", url="http://python.org/"))
     rv = client.get("/admin/")
     data = rv.get_data(as_text=True)
     # print(data)